refactor(round-result): RoundResult에서 OutputView로 클래스 이름 변경
